% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIEXP4.R
\name{CIEXP4}
\alias{CIEXP4}
\title{Gives estimated values of SSI and its confidence bounds}
\usage{
CIEXP4(x1, x2, x3, x4, alpha)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{x4}{is sample from Exp Population 4.}

\item{alpha}{is the level of significance}
}
\description{
Gives estimated values of SSI and its confidence bounds
}
